feat(retranspile): 全量重生成首次归零——Plan 396/B11 上游修复后 rust/src 全绿重建
auto-lang plan-fix/b11-regressions（3f3d0ec3）根治 len() 强转与四类解析缺口后：
- retranspile skip=0（tool.at / error.at 历史上首次在本管线全量转译）
- cargo check Finished（此前 23→7→0）
- rust/src 为全绿重生产物；retranspile.sh 注释同步解锁状态

+++ crates/auto-ai-agent/rust/src/agent.rs
@@ -85,7 +85,7 @@ fn max_tool_result_chars() -> u32 {
 
 pub fn truncate_tool_result(s: &str) -> String {
     let cap = max_tool_result_chars();
-    let count: u32 = ((s.len() as u32) as u32);
+    let count: u32 = ((s.len() as i64) as u32);
     if count <= cap {
         return s.to_string();
     }
@@ -302,7 +302,7 @@ impl Agent {
     }
     pub fn register_skill_tool(&mut self, tool: SkillTool) {
         let block = tool.available_skills_block();
-        match (block.len() as i32) == 0 {
+        match (block.len() as i64) == 0 {
             true => self.skills_block = None,
             false => self.skills_block = Some(block),
         };
